% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{dsrCCA}
\alias{dsrCCA}
\title{Federated rCCA}
\usage{
dsrCCA(logindata, Var, Varx, Vary, lambda1 = 0, lambda2 = 0)
}
\arguments{
\item{logindata}{loginInfo}

\item{Var}{string of the selected colnames to be analyzed}

\item{Varx}{string containing the colnames for the x dataset}

\item{Vary}{string containing the colnames for the y dataset}

\item{lambda1}{regularization term for Cov(x) matrix, default = 0}

\item{lambda2}{regularization term for Cov(y) matrix, default = 0}
}
\value{
cor vector of correlation coefficients, one for each canonical dimension

xcoef coefficient matrix for x dataset

ycoef coefficient matrix for y dataset

cv list containing canonical variates for x dataset (cvx) and y dataset (cvy)

loadings list of loading matrices: load.xx for cor(cvx, x), load.yy for cor(cvy, y),load.xy for cor(cvx, y), load.yx for cor(cvy,x)
}
\description{
Federated rCCA
}
